style(catalog): simplify borders and unify component grouping for cleaner layout
