+++ src/features/catalog/components/common/CompoundTable.tsx
@@ -1,21 +1,14 @@
 "use client";
 import { Table } from "@/components/ui/table";
-import { useCatalogData } from "@/features/catalog/hooks/common/useCatalogData";
-import { useCatalogStore } from "@/features/catalog/store/catalogStore";
 import { useTranslations } from "next-intl";
 import type { ExtendedCompound } from "@/features/catalog/hooks/common/useCompoundData";
-import { useCallback, useState } from "react";
-import { useCompoundColumns } from "./compoundColumns";
-import { useColumnWidths } from "@/features/catalog/hooks/common/useColumnWidths";
 import { TablePagination } from "@/features/catalog/components/common/TablePagination";
 
-import { getCellValue } from "@/features/catalog/utils/getCellValue";
 import { CompoundTableToolbar } from "./CompoundTableToolbar";
 import { CompoundTableHeader } from "./CompoundTableHeader";
 import { CompoundTableRows } from "./CompoundTableRows";
 import { AdvancedFiltersPanel } from "./AdvancedFiltersPanel";
-import { TableColumnKey } from "@/features/catalog/domain/types/TableColumnKey";
-import type { BasicAdvancedFilters } from "@/features/catalog/domain/types/ChemicalCompound";
+import { useCompoundTableHandlers } from "../../hooks/common/useCompoundTableHandlers";
 
 interface CompoundTableProps {
   data: ExtendedCompound[];
@@ -24,7 +17,6 @@ interface CompoundTableProps {
 export function CompoundTable({ data: _data }: CompoundTableProps) {
   const t = useTranslations();
 
-  // Usa o novo hook que integra com o store
   const {
     paginatedData,
     isLoading,
@@ -36,88 +28,23 @@ export function CompoundTable({ data: _data }: CompoundTableProps) {
     selectedCategories,
     sortColumn,
     sortOrder,
-    totalCompounds: _totalCompounds,
-    filteredCount: _filteredCount,
     advancedFilters,
-  } = useCatalogData();
-
-  // Actions do store
-  const {
     setSearchTerm,
     setSelectedCategories,
     setCurrentPage,
     setRowsPerPage: _setRowsPerPage,
-    setSortColumn,
-    toggleSortOrder,
-    setAdvancedFiltersOpen,
-    setAdvancedFilters,
-    resetAdvancedFilters,
     toggleColumn,
     visibleColumns,
-  } = useCatalogStore();
-
-  // Estado local para colunas
-  const [columnsMenuOpen, setColumnsMenuOpen] = useState(false);
-  const allColumns = useCompoundColumns();
-
-  // Função para obter valor da célula
-  const getCellValueForWidth = useCallback(
-    (compound: ExtendedCompound, key: TableColumnKey) => {
-      return getCellValue(compound, key, t);
-    },
-    [t]
-  );
-
-  const columnWidths = useColumnWidths(
-    paginatedData,
+    columnsMenuOpen,
+    setColumnsMenuOpen,
     allColumns,
-    getCellValueForWidth
-  );
-
-  // Handlers
-  const handleSafeSort = (key: TableColumnKey) => {
-    if (
-      typeof key === "string" &&
-      [
-        "id",
-        "name",
-        "commonName",
-        "synonym",
-        "formula",
-        "casNumber",
-        "molarMass",
-        "physicalForm",
-        "meltingPoint",
-        "boilingPoint",
-        "density",
-        "refractiveIndex",
-        "solubility",
-        "solubilityNumeric",
-        "solubilityQualitative",
-      ].includes(key)
-    ) {
-      if (sortColumn === key) {
-        toggleSortOrder();
-      } else {
-        setSortColumn(key as keyof ExtendedCompound);
-      }
-    }
-  };
-
-  const handleAdvancedFiltersChange = useCallback(
-    (filters: BasicAdvancedFilters) => {
-      setAdvancedFilters(filters);
-    },
-    [setAdvancedFilters]
-  );
-
-  const handleAdvancedFiltersReset = useCallback(() => {
-    resetAdvancedFilters();
-  }, [resetAdvancedFilters]);
-
-  const handleAdvancedFiltersToggle = useCallback(() => {
-    setAdvancedFiltersOpen(!advancedFilters.isOpen);
-  }, [advancedFilters.isOpen, setAdvancedFiltersOpen]);
+    getCellValueForWidth,
+    columnWidths,
+    handleSafeSort,
+    handleAdvancedFiltersChange,
+    handleAdvancedFiltersReset,
+    handleAdvancedFiltersToggle,
+  } = useCompoundTableHandlers();
 
   if (isLoading) {
     return (
@@ -140,38 +67,41 @@ export function CompoundTable({ data: _data }: CompoundTableProps) {
   return (
     <div
       className={`
-         w-full max-w-8xl mx-auto my-10 space-y-4 px-4 md:px-8 py-4 border 
-    border-border rounded-lg bg-background shadow-sm 
-    dark:bg-zinc-900 dark:border-zinc-700 
+        w-full max-w-8xl mx-auto my-10 px-4 md:px-8 py-6
+        border border-border dark:border-zinc-700
+        rounded-xl shadow-md bg-background dark:bg-zinc-900
+        space-y-6
       `}
     >
-      {/* Painel de Filtros Avançados */}
-      <AdvancedFiltersPanel
-        filters={advancedFilters.filters}
-        onFiltersChange={handleAdvancedFiltersChange}
-        onReset={handleAdvancedFiltersReset}
-        isOpen={advancedFilters.isOpen}
-        onToggle={handleAdvancedFiltersToggle}
-        isActive={advancedFilters.isActive}
-      />
+      {/* Toolbar + Filtros basicos */}
+      <div className="space-y-4">
+        <CompoundTableToolbar
+          selectedCategories={selectedCategories}
+          setSelectedCategories={setSelectedCategories}
+          searchTerm={searchTerm}
+          setSearchTerm={setSearchTerm}
+          columnsMenuOpen={columnsMenuOpen}
+          setColumnsMenuOpen={setColumnsMenuOpen}
+          allColumns={allColumns}
+          visibleColumns={visibleColumns}
+          toggleColumn={toggleColumn}
+          t={t}
+        />
 
-      {/* Toolbar com busca e filtros básicos */}
-      <CompoundTableToolbar
-        selectedCategories={selectedCategories}
-        setSelectedCategories={setSelectedCategories}
-        searchTerm={searchTerm}
-        setSearchTerm={setSearchTerm}
-        columnsMenuOpen={columnsMenuOpen}
-        setColumnsMenuOpen={setColumnsMenuOpen}
-        allColumns={allColumns}
-        visibleColumns={visibleColumns}
-        toggleColumn={toggleColumn}
-        t={t}
-      />
+        {/* Painel de Filtros Avançados */}
+        <AdvancedFiltersPanel
+          filters={advancedFilters.filters}
+          onFiltersChange={handleAdvancedFiltersChange}
+          onReset={handleAdvancedFiltersReset}
+          isOpen={advancedFilters.isOpen}
+          onToggle={handleAdvancedFiltersToggle}
+          isActive={advancedFilters.isActive}
+        />
+      </div>
 
-      {/* Scroll horizontal só quando necessário */}
-      <div className="w-full overflow-x-auto">
-        <Table className="min-w-full table-fixed shadow-xl">
+      {/* Tabela */}
+      <div className="overflow-x-auto">
+        <Table className="min-w-full table-fixed">
           <CompoundTableHeader
             allColumns={allColumns}
             visibleColumns={visibleColumns}
@@ -193,7 +123,7 @@ export function CompoundTable({ data: _data }: CompoundTableProps) {
       </div>
 
       {/* Paginação */}
-      <div className="flex items-center justify-between mt-4">
+      <div className="flex items-center justify-between pt-2">
         <div className="flex items-center gap-2">
           <span className="text-sm text-muted-foreground">
             {t("compoundTable.rowsPerPage")}
@@ -203,12 +133,14 @@ export function CompoundTable({ data: _data }: CompoundTableProps) {
             onChange={(e) => _setRowsPerPage(Number(e.target.value))}
             className="border rounded px-2 py-1 text-sm bg-background dark:bg-zinc-800 dark:border-zinc-700"
           >
-            <option value={5}>5</option>
-            <option value={10}>10</option>
-            <option value={20}>20</option>
-            <option value={50}>50</option>
+            {[5, 10, 20, 50].map((v) => (
+              <option key={v} value={v}>
+                {v}
+              </option>
+            ))}
           </select>
         </div>
+
         <TablePagination
           currentPage={currentPage}
           totalPages={totalPages}
